Add files via upload
This game is the first game I ever created in Python which explains how one-dimensional and primitive it is :))
In it, the player has to catch one of two apples. The one being a good apple that increases the score by 1 point and another apple(rotten apple) that ends the game when caught.